Refactor startup checks, initial implementation for java version check.
